Comprises of python codes as a part of project flora robotica
1.3altpixie_control-  controls the various pixies at different power
levels from the raspberry pi using config files that are used for
setting the specifications of the pixie. The values of light sensors are
logged to csv files.
2. Final Automate- A code that automates the various process such as
watering the plants, heating and day night cycle through RF Modules. The
RF codes are identified using RF Snipper and passed through raspberry pi
transmitter receiver port
3.Final_tip_detection- A code that uses corner harris algorithm and does
motion tracking of the tip of the plant and records the coordinate
position.
4.sharingcode2.sh-A shell script to mount the local drive of raspberry
pi to the flash drive
5.THE_FINAL-Code for controlling the direction of the motor based on the
light sensor value. Uses multitasking of python to operate the motors
simultaneously.
6.try7-A code that automatically transfers the pictures from the
raspberry pi to backup flash drive every 30 minutes with the help of
advanced python scheduler.